- Moved vacancies view holder and adapter - Implemented correct word endings
