+++ app/src/main/res/values/strings.xml
@@ -10,7 +10,9 @@
     <string name="second_developer">Mysteri027</string>
     <string name="third_developer">kukidevs</string>
     <string name="fourth_developer">vaiwork</string>
-    <string name="num_jobs_found">Найдено %1$d вакансий</string>
+    <string name="num_jobs_found_5_9">Найдено %1$d вакансий</string>
+    <string name="num_jobs_found_1">Найдено %1$d вакансия</string>
+    <string name="num_jobs_found_2_4">Найдено %1$d вакансии</string>
     <string name="no_jobs_found">Таких вакансий нет</string>
     <string name="empty_list">Список пуст</string>
     <string name="apply">Применить</string>
